Be more explicit about allow less precise FS and XS transitions
The goal here is to explicitly allow for implementations that use two
states to track FS.  The previous wording was ambiguous here, as it
seemed to allow that in the text but then specify a set of state
transitions.  This new wording makes it explicit that the listed state
transitions are just the most precise possible, and defines how
implementations are allowed to be less precise.