Don't describe fence.tso as a pseudoinstruction
As discussed in #186, it makes no sense for RISC-V assemblers to accept
`fence.tso rw, rw`. Removing this expansion from the pseudoinstruction
table (23.2) seems consistent with the description of fence.tso
elsewhere in the document.

If different pred/succ for fence.tso are made legal in a future spec, at
that point it would make sense to list fence.tso as a pseudoinstruction.