Un-mix trap and exception in {m,s}epc
This might be a bit pedantic, but the current wording in the spec mixes
up traps and execptions here in a way that makes it sound like
interrupts don't write {m,s}epc if you're reading quickly.

I don't really like the wording I've proposed, so this is more of a bug
report than a fix.