"c.addi x0, 0" isn't a legal instruction
I was reviewing some QEMU patches recently and noticed an oddity of the
current description of the ISA manual: the NOP instruction can also
decode as a valid ADDI instruction, but the C.NOP instruction cannot
decode as a valid C.ADDI instruction despite the ISA manual saying
"C.NOP is encoded as c.addi x0, 0 and so expands to addi x0, x0, 0".
This isn't a big deal, because "c.nop" does match what you'd expect the
encoding of "c.addi x0, 0", but it is technically incorrect because the
instruction "c.addi x0, 0" doesn't exist.

This patch fixes the text of "c.nop" to just say it expands to "NOP".

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>