Indicate that mcycle can be shared between harts
When reading through the manual for the QEMU patches to implement
mcounterinhibit I found that the commentary allows mcycle to be shared
between harts.  I actually really don't like this behavior, but I can
understand that it may make sense for some systems.  Regardless, the
normative text of the ISA manual should match the commentary.

This patch introduces the concept that modifications of a CSR on one
hart may cause the value on other harts to change.  If I understand
correctly that's not quite a WARL field, which can only change upon
writes.  I don't think it's worth giving this a name, though, as
hopefull it'll only apply to mcycle.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>