Don't mandate that multiple harts on a core share mcycle
The commentary indicates that sharing mcycle between harts is mandatory,
but I don't think that's a good idea -- both because it's introducing a
constraint that is very hard to formalize (ie, WTF is a core) and
because I think some systems will desire different behavior.

This patch softens the wording in the commentary to avoid the appearance
that sharing mcycle between threads is mandatory.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>