Refactor LR/SC reservation set terminology
Define invalidations in terms of the reservation itself,
defining the invalidation set simply as a mechanism that
expands the size of reservations. This keeps definitions
independent of whether a reservation might span multiple
reservation sets, as is unavoidable with unaligned LR/SC.
A minimum sized reservation set allows exact reservation.

Reservations being separate from the bytes they refer to,
introduce the term "reserved byte" to refer to the bytes.