ignore write to "controlled" SBE and UBE.
Behaviour when endian fields are hard-coupled is not well defined.
If both fields are written with opposite endianness which takes precedence,
 or are the fields left unchanged?
Both approaches are currently used in CSRs.
Further complicating the situation on RV32 systems is the separation
of UBE from MBE/SBE into two CSRs.
Whereas precedence and conflict cannot be determined for two discrete CSR
 updates I propose instead that controlled fields not be directly writable.

Signed-off-by: David Horner <ds2horner@gmail.com>