Add virtual instruction exceptions to the hypervisor extension
For the hypervisor extension, add another exception type, "virtual instruction exception", that substitutes for the illegal instruction exception in cases where a hypervisor is expected to emulate the instruction.  This saves time because this exception type can be delegated directly to HS mode, whereas illegal instructions will typically be processed first in M mode and delegated to HS mode only by software.