Clarifications to the perceived ordering of CSR accesses and their effects
More clearly distinguished the ordering rules that apply to the same hart from those that involve other agents (harts and devices).  Made CSR side effects completely synchronous to the same hart by default.  (Deviances are still allowed, when documented.)  Added a comment that rules about the reordering of CSR accesses in the global memory order should probably be moved to the RVWMO chapter.