Add an "lga" pseudoinstruction for GOT loads even in non-PIC
This is useful for cases where compilers and/or assembly writers want to
force the use of a GOT. For example, this could be used to address [1],
but more generally it's a useful thing to have, and is a strange hole in
the set of pseudoinstructions provided.

Whilst here, alter the description of "la" to not mention absolute. All
addresses are absolute, but highlighting it for "la" specifically is
confusing as it potentially implies that it's using an absolute lui/addi
sequence rather than a PC-relative sequence.

[1] https://github.com/riscv/riscv-elf-psabi-doc/issues/126