Improve description of interrupt traps
Simplify the description of when interrupt traps are taken, in part by
abandoning the concept that interrupt sources are globally enabled
selectively, depending on delegation registers.  (Global enablement on
a per-interrupt basis had always been in conflict with other text in
the document that says that interrupts are globally enabled for M-mode
or S-mode simply when field MIE or SIE in mstatus is set.)

Also make clear that pending-and-enabled interrupts must trap
immediately, before another instruction is executed.  This requirement
is sufficient to prove (within the bounds of what is enforceable) that
interrupts must have higher priority than synchronous exception traps.