Remove the "C" bit, per virt mem TG vote
We may choose to re-add "C" later, but for now it was decided that
reserving a bit for "C" was premature.

Also:

Fix Sv57, which somehow still had a four-level PPN diagram

Fix a sentence saying that setting PTE reserved bits should trigger an
access fault.  Any incorrect use of PTEs should be a page fault.