Remove L placeholder chapter from spec
Since the L placeholder chapter was drafted (over 10 years ago!), there
has been almost no detectable interest, to the point that not even a list
of instructions was ever proposed for the L extension.

If decimal FP becomes important at some point in the future, we can define
an ISA extension to accelerate it at that time.