Fix editing error in mtval/stval definition
Partially introduced by 7f45777bfc3e0a6b327552a286392ccfd5c933a8 and 94ab414c4b470edae8bde4e7759163fe7ddf8ab4

The spec says that the platform will define which exceptions must populate
mtval or stval with a nonzero value and which are permitted to zero it.
But then later statements seem not to permit the possibility of zeroing
the register, which is a contradiction.  Resolve in a backwards-compatible
manner by permitting zeros unless the platform says otherwise (which was
the original intent, anyway).

This permits e.g. zeroing mtval on an EBREAK instruction, where mepc
provides sufficient information, as was the intent of 94ab414c4b470edae8bde4e7759163fe7ddf8ab4.