Add Hazard3 to open-source marchid list (#784)
This is a 3-stage `RV32I` core, with optional support for `M`/`C`/`Zba`/`Zbb`/`Zbc`/`Zbs` and debug support. It passes the ISA compliance tests, riscv-formal, OpenOCD DM compliance tests and the end-to-end debug tests from riscv-tests/debug.

[Github](https://github.com/Wren6991/Hazard3/blob/master/License)

[PDF documentation](https://github.com/Wren6991/Hazard3/blob/master/doc/hazard3.pdf)

The license is DWTFPLv3, which is effectively a public domain dedication.