Cleanup build errors and remove generated PDF
The PDF removed here is generated by the make command and will be added
as an artifact to a GitHub action rather than checked in here.

Several build errors existed with regards to missing wavedrom files.
Comment them out for now until they can either be generated or manually
created.

Signed-off-by: Stephano Cetola <scetola@linuxfoundation.org>