satp.mode is not WARL
satp.mode was described as WARL in the table, but has specific behavior
defined by the text of the specification: it requires writes to the
whole CSR to be avoided when the MODE field is invalid (and WARL is
defined on fields, not CSRs).

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>

---

I don't really care which behavior is desired, but my guess is that the
more verbose language ended up there on purpose.