Clarify that LR/SC reservation set restriction is on physical pages
As the A extension already indicates, an SC might succeed on some
implementations even if it uses a virtual address on a different page
as the LR instruction reserved, so long as the underlying physical
pages are the same.  So, clarify that the restriction on reservation
set dimensions refers to the underlying physical page.

Resolves #966