Disambiguate the use of the term "reserved" for HINT instructions
As was raised in #300, the overloaded use of the term "reserved" when
listing HINT encodings is potentially confusing. As clarified in that
thread, these encodings are "reserved" in the intuitive English language
meaning of the word, but are _not_ part of the "reserved instruction-set
category" as defined in the section 1.3 introduction and further
elaborated in section 2.2 "Base Instruction Formats" (which notes how
platforms can choose to raise an illegal-instruction exception for
reserved instruction encodings).