Fixes for atomics wording
Fix L.aq and S.rl comment:
The aq and rl bits are not both necessary for C++ seq_cst semantics.
See table A.7

seq_cst LR/SC:
The SC.aqrl recommendation here was inconsistent with the mapping in A.6
and unnecessary for A.7. It is probably correct for a trailing fence
mapping, but that's not discussed anywhere.

Tweak load-acquire and AMO sequential consistency wording.

Demote this discussion to commentary.

Movea the last figure up to prevent it from appearing in the commentary,
which would have looked ugly to me.

A.6 and A.7 discussion:
Note that A.6 and A.7 are incompatible as is, and there is no smooth
transition from A.6 to A.7