Fixes for Atomics wording
Fix L.aq and S.rl comment:
The aq and rl bits are not both necessary for C++ seq_cst semantics.
See table 55 (formerly A.7).

seq_cst LR/SC:
The SC.aqrl recommendation here was inconsistent with the mapping in
54 (A.6) and unnecessary for 55 (A.7). It is probably correct for a
trailing fence mapping, but that's not discussed anywhere.

Tweak load-acquire and AMO sequential consistency wording.

Demote this discussion to commentary.

A.6 and A.7 (now 54 and 55) discussion:
Note that 54 (A.6) and 55 (A.7) are incompatible as is, and there is no
smooth transition from 54 to 55.