Update hypervisor.adoc
I think the fix here is to move the backtick outside of the apostrophe s to make the whole mtval's monospace.  Alternatively, we can surround mtval with double backticks, such as ``mtval``'s but that results in the apostrophe s being spaced slightly beyond the word mtval which looks odd.