Explicitly allow side effects for a failed SC
Since PTE updates are specified to occur as a result of "memory
accesses" it is possible to interpret the old wording as requiring no
PTE update for a failed SC, since there is no memory access.  However,
the PTE update is part of the translation process, which on natural
implementations will occur before the reservation validity check can
occur.

This is consistent with the wording in "Addressing and Memory
Protection" which allows traps to prevent the memory access without
preventing the PTE update.