Opcode 0x77 is named OP-VE, for "Vector Extensions"
Resolves #1310

This major opcode is currently used by vector crypto extensions.  It'll
likely be used by future vector extensions, as well.

The official naming table currently describes this opcode as _reserved_,
which is obviously no longer the case.  Informally, it has sometimes been
referred to as OP-P, since there was once a plan to encode the packed-SIMD
extensions there.  The vector crypto spec reflects this informal usage;
this commit fixes that up.  (The plan to use OP-P for packed SIMD has been
superseded by a proposal to use brownfield space elsewhere.)