Sstc editorial fixes. (#1315)
* Sstc editorial fixes.

The second to last paragraph starts "When STCE in menvcfg is zero, ...".  This should be clarified to say "When this extension is implemented and STCE in menvcfg is zero, ...".  And ditto for the beginning of the last paragraph.

* Rewording final paragraph for STCE clarity.

Change wording to "But when STCE is menvcfg is one and STCE in henvcfg is zero".