Added support for EPUB3
This pull requests adds support for building EPUB3 versions of the
RISC-V ISA manuals. To achieve this, `asciidoctor-epub3` extension has
been used.

Known issues: some eBook readers are not capable of rendering embedded
images. The solution to this problem is disabling embedded image
altogether. The `README.md` file provides more details on how to achieve
this.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan-cristian.cirstea@tutanota.com>