Run docker as the current user
If you have real Docker (not Podman) then by default it will run everything as root, so your build directory is owned by root and the Makefile will not work at all because it tries to modify the build directory.

This adds a flag so that Asciidoctor runs as the current user in Docker instead of root.