Update description of mtval2 in CSR table
While it is always a guest physical address for the hypervisor
traps, when used by a double trap it holds an mcause value.

Change the description to "second trap value" to match the description
used in the section heading. While touching this line also sync the
description of the other CSRs in this part of the table as well as the
*tval registers for the other privilege levels.

Signed-off-by: Alex Richardson <alexrichardson@google.com>