Run root Docker as the current user
Reapplication of 3539eff338352aba83dbae9e7b310d6a7d9a7d56 but this time we only add the `--user` argument for root Docker. Rootless Docker behaves the same as Podman and doesn't need the `--user` flag; in fact it breaks builds if you add it.

I also added a comment for the `:z` flag.