Clarify note that compilers may need to change vtype for vmv<nr>r.v
In a similar vein to 856fe5b, this note states that compilers don't
need to "know or change vtype" for whole vector register moves.

There's some truth to this in that compilers can largely ignore SEW
and LMUL, but ultimately they do depend on vtype and the current
wording might be misleading. A compiler may in fact need to change
vtype to clear vill, see:
https://github.com/llvm/llvm-project/issues/114518