Explain `sscratch` is only swapped by software
Someone familiar with the interrupt handlers of other ISAs may
incorrectly imagine that the current description means that RISCV cores
will automatically perform the swap of `sscratch` with a user register.

This is not the case: it is a recommendation disguised as a description.
Flesh it out a bit so it is less ambiguous.