a-st-ext: remove manual page breaks
We should improve the tooling instead of adding manual page breaks.

Also replace "above" with a reference, because "above" is not correct
anymore and most importantly, it's not acceptable for a spec.

This does not imply any acceptability of the the rest of the sentence,
I just didn't want to leave the spec in a worse shape.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>