Add basic support for tagging text in the spec
This adds two things:

1. Custom CSS to add a subtle dotted underline on hover to any element with an ID starting with `qx_` (an arbitrary prefix unlikely to be used by chance).
2. An example tag `qx_x0_is_zero`, tagging the text that specifies that x0 is hardwired to 0.

This just adds a single tag as an exmaple, but the intention is that such tags would be added throughout the spec, allowing coverage, test plans, tests, documentation, etc. to all link to specific parts of the spec.

The text of these tags can also be extracted via Asciidoc's docbook output format, enabling more use cases such as monitoring for spec changes.