Clarify description of constrained LR/SC loops for Zca++
This PR fixes an editing error in which "C" should have been changed to
"Zca" when the various Zc extensions were incorporated.  It also clarifies
that any compressed instruction, not just those in Zca, that map to one of the
permitted "I" instructions can be used in a constrained LR/SC loop.  This
was always the case, but it was stated obliquely.

Resolves #1938