Remove [5:4] after spimm in cm.push/pop wavedrom and assembly syntax explanation.
The text refers to spimm as "the number of additional 16-byte address increments allocated for the stack frame."
and "The total stack adjustment represents the total size of the stack frame, which is stack_adj_base added
to spimm scaled by 16"

The "Example RV32I PUSH/POP sequences" show spimm with values of 0-3.

Based on those sections, I think spimm is a 2-bit value rather than a 6-bit value.