Avoid describing fixed-point RDN rounding mode as truncation
In the fixed-point context, describing RDN as truncation is correct, since the
effect of lopping off the LSBs is to round towards -inf for both unsigned and
two's complement integers.

Because truncation effects a rounding towards zero in other number formats,
including IEEE 754, truncation has become conflated with rounding towards
zero.  (The terminology even shows up in the C standard.)
This is unfortunate, but it is an avoidable source of confusion.