fix: mstateen*h and hstateen*h access upper half of 64-bit registers
For `mstateen*` and `hstateen*`, the spec states:
> For RV32, the registers listed above are 32-bit, and for the machine-level and
> hypervisor CSRs there is a corresponding set of high-half CSRs for the upper 32
> bits of each register

This is self-contradictory. If "the registers listed above are 32-bit", there
are no "upper 32 bits" to access.

From the discussion in issue #1255, there is general agreement that CSRs that
are defined as 64-bit and have "upper-half" analogues to be used for RV32 are
indeed architecturally 64-bit.

Remove the assertion that the 64-bit registers are 32-bit for RV32.

(Also related to issue #1966.)

Signed-off-by: Paul A. Clarke <pclarke@ventanamicro.com>