Avoid stating that G bit is reserved in PTE of G-stage (#2155)
Avoid stating that the  G bit is reserved in the PTE of the G-stage address translation since the spec requires the G bit to be ignored.  Otherwise, we have a contradiction with an earlier section of the spec which states that a reserved bit in a PTE would lead to a page fault when such bit is set.

Signed-off-by: Joseph Rahmeh <jrahmeh@tenstorrent.com>