Zihintntl compressed hints are predicated on Zca, not C
This fixes an editing error that arose because the Zc* extensions were
merged after Zihintntl, but Zihintntl was not updated accordingly.

Ordinarily, this kind of change would be incompatible, on the basis that
we would be shoehorning instructions into Zca after the fact.  But these
are HINTs, and so functionally they must be present on all implementations
of Zca, regardless of the implementer's intent to implement Zihintntl.
Hence, no HW implementation is invalidated by making this change.

Resolves #2212